Addition of Notes and Mods
I have included ASM mods that can be added to the game. The level skip mod is still, techincally, in progress, as I have not found a suitable location in ROM to insert it that won't break the game in some way.

The note files are to keep track of ideas, concepts, and experimets that relate to game code and locations in RAM. Eventually, I hope to have a complete RAM map put together.